Split test, create commons.py for tests
