+++ tests/commons.py
@@ -1,0 +1,2 @@
+def build_timestamp(year: int, month: int, day: int):
+    return (year, month, day, 17, 3, 38, 1, 48, 0)
