+++ tests/get_all_entites_from_n_days_test.py
@@ -1,0 +1,6 @@
+import unittest
+
+class TestAllEntitiesFromNDays(unittest.TestCase):
+
+    def test_x(self):
+        pass
